location-refactoring
