+++ src/App.css
@@ -36,10 +36,3 @@
     transform: rotate(360deg);
   }
 }
-
-a#urlButton {
-  color: #61dafb;
-}
-a:-webkit-any-link {
-  color: beige;
-}
