+++ src/components/Show.js
@@ -5,7 +5,7 @@ import GridList from '@material-ui/core/GridList'
 import GridListTile from '@material-ui/core/GridListTile'
 import GridListTileBar from '@material-ui/core/GridListTileBar'
 import Tooltip from '@material-ui/core/Tooltip'
-import { Button } from '@material-ui/core'
+import Typography from '@material-ui/core/Typography';
 import { IconButton } from '@material-ui/core'
 
 const useStyles = makeStyles((theme) => ({
@@ -26,6 +26,11 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
+const MyComponent = React.forwardRef(function MyComponent(props, ref) {
+  // console.log(props)
+  return <div {...props} ref={ref}></div>
+});
+
 const Show = (props) => {
   const shop = props.shops
   // console.log(shop)
@@ -36,26 +41,25 @@ const Show = (props) => {
     <div className={classes.root}>
       <GridList cellHeight={150} className={classes.gridList} spacing={8} cols={3}>
         {shop.map((shop, index) => (
-          <GridListTile key={shop.photo.pc.l} >
-            <img 
-              srcSet={shop.photo.pc.l}
-              alt={shop.name}
-            />
-            <Tooltip title={shop.name} aria-label={shop.name}>
-            <GridListTileBar
-                title={shop.name}
-                actionIcon={
-                  <IconButton aria-label="詳細" className={classes.icon}>
-                    {/* <Button > */}
-                    <a id="urlButton" href={shop.urls.pc} target="_blank">
-                    </a>
-                    {/* </Button> */}
-                  </IconButton>
-                }
-              />
-              </Tooltip>
-                {/* <a href={shop.urls.pc} target="_balnk">hotpeperで開く</a> */}
-            </GridListTile>
+          <GridListTile key={shop.photo.pc.m} >
+            <img srcSet={shop.photo.pc.l} alt={shop.name} />
+            <Tooltip title={shop.name} placement="top-end">
+              <MyComponent>
+                <GridListTileBar
+                  title={shop.name}
+                  actionIcon={
+                    <IconButton >
+                      <Typography >
+                        <a id="urlButton" href={shop.urls.pc} target="_blank" rel="noopener noreferrer" className={classes.icon}>
+                          info
+                        </a>
+                        </Typography>
+                    </IconButton>
+                  }
+                />
+              </MyComponent>
+            </Tooltip>
+          </GridListTile>
         ))}
       </GridList>
       {/* <ul>
